moving vuex store for start of vuex pattern
