+++ src/main.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import App from "./App.vue";
 import router from "./router";
-import store from "./store";
+import store from "./store/store";
 import UserConfigService from "@/services/services/UserConfigurationService";
 import SessionService from "@/services/services/SessionService";
 import Util from "@/lib/util";
